Cambios productos controller

+++ src/main/java/com/utec/farmacia/dao/productos/ProductosModel.java
@@ -13,7 +13,7 @@ public class ProductosModel {
     private int idproducto;
     private String nombre;
     private String descripcion;
-    private double precio;
+    private float precio;
     private int existencia;
     private String estado;
     private int codigo;
@@ -44,11 +44,11 @@ public class ProductosModel {
         this.descripcion = descripcion;
     }
 
-    public double getPrecio() {
+    public float getPrecio() {
         return precio;
     }
 
-    public void setPrecio(double precio) {
+    public void setPrecio(float precio) {
         this.precio = precio;
     }
 
@@ -93,8 +93,6 @@ public class ProductosModel {
     }
 
 
-   
-    
             
      
 }
